requirements.txt
-e .
pyro-ppl==1.6.0
matplotlib==3.4.1
